Add in a new commands concept so that we begin to have real structure to hang stuff on.
In particular, we add a new "Command" object, a "Commands" module to start adding commands to, and we start the main processing loop in "main".

This doesn't test much because this aims mostly to setup stuff around "Main" instead of the other files - "addsongs" for instance hasn't been battle hardened yet, and thus doesn't really need testing.