Overhaul Oracles so they distinguish between "current" and "next" songs.
You need a way to get the currently playing song in some instances, but "state" is a finnicky beast. As a result, this jumps through a whole buncha hoops to disentangle the concept of "next" and "current", to lay out expectations of behavior for the two that a standard human might have, and to ensure the code's logic adheres to those expectations.

We also begin making some guarantees around the interface, such as "current_song" will never change it's return value until "next_song" is called. This has ripple effects, but should make it easier for potential users and integrators to wrap their heads around what's going on.