Change loading media libraries to be a "copy" event rather than a "creation" event.
If you load a library, and then replace the object with a different object, you won't see the new object if you're passing things around by pointers. So,  use "copy" idioms instead of overwriting the pointer in just one spot.

This codebase is in serious need of some love though...