Add new "listplaylists" and "listsongs" commands to the API.
Still need to be tested and what not, but I've verified "listplaylists" works. Next up is to verify "listplaylists" and "listsongs" locally, then write the tool for "listsongs". I'm a little worried the "listplaylists" entry is lacking some stuff that's necessary, and I'm also worried that I'm underscoping that API with the overarching plans I have, but it's better to hammer in small functionality now that you can build on, than shoot for the moon.