Create command_types_v1.py as a stub API specification for the new websockets based implementation.
I decided after evaluating Javascript + RPCs that it was more of a pain than it was worth. GRPC is difficult to install except under the best of circumstances (Linux, which I'm not developing on since most users won't be using Linux) and distribution in Javascript files is difficult to say the least.

Websocket implementations are broadly available, do 90% of the work of GRPC, and the remaining 10% isn't worth the pain for a project this size - where we have small numbers of machines talking to each other in a limited capacity using highly debuggable languages in a performance-insensitive context, and where we care much more about browser-debuggability, compatibility, and readability than we care about performance or backend-concerns.

Too bad - I really like the strongly defined schemas of protos, as well as the strong transpilation support so moving between languages is damn easy and requires little thought =/.