Replace the stubby little "flags" library with absl.flags
I was hoping it wouldn't be necessary, but it was getting annoying to have to deal with flags logic. This'll make it easier to deal with flags throughout the codebase.

Tests which want to remain compatible with Pycharm's test runner will need to subclass "common.test_utils.FlagsTest" in order for flags to remain available.