Create an event-based API from the ashes of the Command/response API.
New API clients are expected to separate their logic loop into one that
delivers commands, and one that receives events.

Events may be driven by commands, but that is not always the case.

The API was designed this way because it's really, really hard to turn
Websockets into a command/response framework without a lot of groaty
middleware in between. Meanwhile, it's pretty easy to do something
event-driven, where neither side can expect to recieve something in
response to something they sent.

This also takes some time to formalize the Message and MessageMap class,
from the Event/Command stuff I was working on previously. There are a
lot more utility functions and all the logic should make a lot more
sense.

This whole commit violates the "small commit" rule, but it was pretty
difficult to split this out into meaningful, small commits without
wasting a whole bunch of time, so future programmers get to weep at my
impatience I guess.