Reimplement everything using Pydantic. And other stuff.
I don't even care any more - this is the second massive, major refactor
of the same damn system. It's late, I'm tired, and I want this thing in
so I can be happy when I start working tomorrow.

On the upside, I was able to toss out **so much code** and make **so
many things** less-ugly. Likewise, I chewed through most of the project
auto-inspection errors, and finally got around to re-kajiggering the API
tool.

On the downside, I am checking in a lot of stuff that will bite me
later. Probably. I gave it a once-over, but this is a huge commit and I
did not bother with much new testing.